net: extract dial_single helper into net/dial
Move the connect-with-timeout helper out of transport.cc's anonymous
namespace into net::detail::dial_single, declared in a new dial.h and
defined in dial.cc. External linkage prepares it for use by dialing
strategies implemented as templates in the header, which are
instantiated in other translation units.
