+++ src/v/net/BUILD
@@ -32,6 +32,7 @@ redpanda_cc_library(
         "batched_output_stream.cc",
         "conn_quota.cc",
         "connection.cc",
+        "dial.cc",
         "dns.cc",
         "probes.cc",
         "server.cc",
@@ -44,6 +45,7 @@ redpanda_cc_library(
         "connection.h",
         "connection_rate.h",
         "connection_rate_counter.h",
+        "dial.h",
         "dns.h",
         "exceptions.h",
         "fwd.h",
